requirements.txt
streamlit
yt-dlp
openai
gtts
Pillow
requests
python-dotenv